+++ lib/pocketbase.dart
@@ -6,13 +6,12 @@ library pocketbase;
 export "src/auth_store.dart";
 export "src/client.dart";
 export "src/client_exception.dart";
-
 // dtos
 export "src/dtos/admin_auth.dart";
-export "src/dtos/admin_model.dart";
 export "src/dtos/apple_client_secret.dart";
 export "src/dtos/auth_method_provider.dart";
 export "src/dtos/auth_methods_list.dart";
+export "src/dtos/auth_model.dart";
 export "src/dtos/collection_model.dart";
 export "src/dtos/external_auth_model.dart";
 export "src/dtos/health_check.dart";
@@ -20,7 +19,6 @@ export "src/dtos/jsonable.dart";
 export "src/dtos/log_request_model.dart";
 export "src/dtos/log_request_stat.dart";
 export "src/dtos/record_auth.dart";
-export "src/dtos/record_model.dart";
 export "src/dtos/record_subscription_event.dart";
 export "src/dtos/result_list.dart";
 export "src/dtos/schema_field.dart";