+++ lib/src/auth_store.dart
@@ -1,12 +1,14 @@
 import "dart:async";
 import "dart:convert";
 
+import "dtos/auth_model.dart";
+
 /// Event object that holds an AuthStore state.
 class AuthStoreEvent {
   AuthStoreEvent(this.token, this.model);
 
   final String token;
-  final dynamic /* RecordModel|AdminModel|null */ model;
+  final AuthModel? model;
 
   @override
   String toString() => "token: $token\nmodel: $model";
@@ -16,14 +18,14 @@ class AuthStoreEvent {
 /// the authenticated User/Admin model and its token.
 class AuthStore {
   String _token = "";
-  dynamic /* RecordModel|AdminModel|null */ _model;
+  AuthModel? _model;
   final _onChangeController = StreamController<AuthStoreEvent>.broadcast();
 
   /// Returns the saved auth token (if any).
   String get token => _token;
 
   /// Returns the saved auth model (if any).
-  dynamic /* RecordModel|AdminModel|null */ get model => _model;
+  AuthModel? /* RecordModel|AdminModel|null */ get model => _model;
 
   /// Stream that gets triggered on each auth store change
   /// (aka. on [save()] and [clear()] call).
@@ -51,7 +53,7 @@ class AuthStore {
   /// Saves the provided [newToken] and [newModel] auth data into the store.
   void save(
     String newToken,
-    dynamic /* RecordModel|AdminModel|null */ newModel,
+    AuthModel? newModel,
   ) {
     _token = newToken;
     _model = newModel;